Espresso model starts working!!!
There is a bit of a snafu here, that needs to be corrected:

The model initialization needed to be moved head of 'sim_post_argv_init'
in order to have HW unit creation in the right place, but that only works,
if the model is specified on the command line. If not, we'll have
to do the init right where we set the callbacks, because 'sd'
member of the scpu object is not filled in earlier. Sad, it's quite
an awful hack. Alas, at least it works now.